remove redundant files
